Apply Rector recommendations across src/ and tests/
Rector's dry-run was flagging 25 files. Most fixes are mechanical
simplifications surfaced by the upgraded ruleset:

- ->where('col', null)  ->  ->where('col')
  Builder treats both forms the same; the second is the canonical
  shorthand. Affects DeviceSessionModel, AccessTokenRepository,
  UserIdentityModel and related queries.
- Drop redundant `instanceof X ? : null` casts when the static type
  already proves the variable is null|X.
- Drop string casts in `assertStringContainsString((string) $needle, ...)`
  when the needle is already a string-typed expression.
- Replace `$this->createMock(Foo::class)` with `Foo::createStub(…)` where
  no behaviour mocking is involved (CreateStubOverCreateMockArgRector).
- Drop unused `MockObject` import flagged by no_unused_imports.

Also reorders imports across 6 test files to match the project's
ordered_imports rule (php-cs-fixer 3.95.1 reapplied automatically).

PHPUnit, PHPStan and Deptrac all still pass; PHPUnit assertion count went
from 1362 to 1373 (Rector's stub conversions register more assertions).

+++ src/Authentication/Passwords/HistoryValidator.php
@@ -35,7 +35,7 @@ class HistoryValidator extends BaseValidator implements PasswordValidatorInterfa
 {
     public function check(string $password, ?User $user = null): Result
     {
-        $retain = (int) ($this->config->passwordHistorySize ?? 0);
+        $retain = $this->config->passwordHistorySize ?? 0;
 
         if ($retain <= 0 || ! $user instanceof User || $user->id === null) {
             return new Result(['success' => true]);
